+++ packages/api-client/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [26.5.0](https://github.com/wireapp/wire-web-packages/compare/@wireapp/api-client@26.4.0...@wireapp/api-client@26.5.0) (2023-10-24)
+
+### Features
+
+* check if removal key is available on backend ([#5662](https://github.com/wireapp/wire-web-packages/issues/5662)) ([441762d](https://github.com/wireapp/wire-web-packages/commit/441762d83a620c253650e8425bfe60b03a848265))
+
+# [26.4.0](https://github.com/wireapp/wire-web-packages/compare/@wireapp/api-client@26.3.3...@wireapp/api-client@26.4.0) (2023-10-17)
+
+### Features
+
+* add ability to blacklist conversations ([#5633](https://github.com/wireapp/wire-web-packages/issues/5633)) ([4e82b87](https://github.com/wireapp/wire-web-packages/commit/4e82b87d62ccbaa8de54ceefa7c60b7f7913e7b3))
+
+## [26.3.3](https://github.com/wireapp/wire-web-packages/compare/@wireapp/api-client@26.3.2...@wireapp/api-client@26.3.3) (2023-10-16)
+
+**Note:** Version bump only for package @wireapp/api-client
+
+## [26.3.2](https://github.com/wireapp/wire-web-packages/compare/@wireapp/api-client@26.3.1...@wireapp/api-client@26.3.2) (2023-10-16)
+
+**Note:** Version bump only for package @wireapp/api-client
+
 ## [26.3.1](https://github.com/wireapp/wire-web-packages/compare/@wireapp/api-client@26.3.0...@wireapp/api-client@26.3.1) (2023-10-11)
 
 **Note:** Version bump only for package @wireapp/api-client